// src/proto/todo.ts
// Code generated by protoc-gen-ts_proto. DO NOT EDIT.
// versions:
//   protoc-gen-ts_proto  v2.6.1
//   protoc               v5.29.3
// source: src/proto/todo.proto

/* eslint-disable */
import { GrpcMethod, GrpcStreamMethod } from "@nestjs/microservices";
import { Observable } from "rxjs";

export const protobufPackage = "todo";

export interface Empty {
}

export interface Todo {
  id: string;
  title: string;
  completed: boolean;
}

export interface TodoList {
  todos: Todo[];
}

export interface TodoById {
  id: string;
}

export interface CreateTodoDto {
  title: string;
}

export interface UpdateTodoDto {
  id: string;
  title: string;
  completed: boolean;
}

export const TODO_PACKAGE_NAME = "todo";

export interface TodoServiceClient {
  findAll(request: Empty): Observable<TodoList>;

  findOne(request: TodoById): Observable<Todo>;

  create(request: CreateTodoDto): Observable<Todo>;

  update(request: UpdateTodoDto): Observable<Todo>;

  remove(request: TodoById): Observable<Empty>;
}

export interface TodoServiceController {
  findAll(request: Empty): Promise<TodoList> | Observable<TodoList> | TodoList;

  findOne(request: TodoById): Promise<Todo> | Observable<Todo> | Todo;

  create(request: CreateTodoDto): Promise<Todo> | Observable<Todo> | Todo;

  update(request: UpdateTodoDto): Promise<Todo> | Observable<Todo> | Todo;

  remove(request: TodoById): Promise<Empty> | Observable<Empty> | Empty;
}

export function TodoServiceControllerMethods() {
  return function (constructor: Function) {
    const grpcMethods: string[] = ["findAll", "findOne", "create", "update", "remove"];
    for (const method of grpcMethods) {
      const descriptor: any = Reflect.getOwnPropertyDescriptor(constructor.prototype, method);
      GrpcMethod("TodoService", method)(constructor.prototype[method], method, descriptor);
    }
    const grpcStreamMethods: string[] = [];
    for (const method of grpcStreamMethods) {
      const descriptor: any = Reflect.getOwnPropertyDescriptor(constructor.prototype, method);
      GrpcStreamMethod("TodoService", method)(constructor.prototype[method], method, descriptor);
    }
  };
}

export const TODO_SERVICE_NAME = "TodoService";